Moved sound from bat cs to new bat sprite cs
