+++ Incorpreal/bat/Bat.cs
@@ -4,7 +4,6 @@ using System;
 public class Bat : KinematicBody2D
 {
     [Export]
-    public float time=0;
     public int moveSpeed = 150;
     public int attack;
     public int health;
@@ -12,7 +11,6 @@ public class Bat : KinematicBody2D
     public string enemyName;
     public GlobalPlayer gp;
     public KinematicBody2D player;
-    public AudioStreamPlayer2D footsteps = new AudioStreamPlayer2D();
 
     public override void _PhysicsProcess(float delta)
     {
@@ -25,14 +23,6 @@ public class Bat : KinematicBody2D
 
     public override void _Ready()
     {
-        this.AddChild(footsteps);
-        const string Path = "res://sounds/BatSound.wav";
-        AudioStream footstep = (AudioStream)GD.Load(Path);
-        footsteps.Stream = footstep;
-        footsteps.Autoplay = true;
-        footsteps.MaxDistance = 300;
-        footsteps.Attenuation = 3;
-        footsteps.VolumeDb = (1);
         var filePath = "res://Enemies/enemies.txt";
         File newFile = new File();
         newFile.Open(filePath, File.ModeFlags.Read);
@@ -77,7 +67,6 @@ public class Bat : KinematicBody2D
             tq.GetChild(1).Name = enemyName;
             tq.GetChild(1).Call("_Ready");
             GetTree().ChangeScene("res://Battle.tscn");
-            footsteps.Stop();
         }
     }
 
@@ -96,18 +85,4 @@ public class Bat : KinematicBody2D
             player = null;
         }
     }
-    public override void _Process(float delta)
-  {
-        time += delta;
-        if (footsteps.Playing == false)
-        {
-            if (time > 5)
-            {
-                time = 0;
-                footsteps.Play();
-            }
-        }
-
-  }
-
 }
